pyproject.toml
[project]
name = "arc"
version = "1.0.0"
description = ""
authors = [
    {name = "Triphase", email = "leminaw@gmail.com"},
]
dependencies = [
    "django>=4.2.7",
    "djangorestframework>=3.14.0",
    "jsonschema>=4.19.2",
    "django-filter>=23.5",
    "django-cors-headers>=4.3.1",
    "Pillow>=10.1.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.1.4",
]

[tool.pdm.options]
add = ["--no-self"]
install = ["--no-self"]
remove = ["--no-self"]

[tool.ruff]
extend-exclude = [
    "*/migrations/*.*",
    ".venv/*.*",
]
select = [
    "F", "E", "W", "I", "N", "UP", "B", "C4", "DJ", "INP", "PIE", "PYI",
    "RSE", "RET", "SLOT", "SIM", "TID", "INT", "PTH", "PERF", "RUF"
]
ignore = ["RUF012"]

[tool.ruff.per-file-ignores]
"arc/settings.py" = ["E501"]
"admin.py" = ["DJ007"]

[tool.ruff.isort]
lines-after-imports = 2
